+++ src/aeolus/plot/mpl.py
@@ -1,11 +1,29 @@
 # -*- coding: utf-8 -*-
 """Matplotlib-related utilities."""
+from typing import Sequence, Union, Optional
+
+from iris.cube import Cube
+from matplotlib.axes._axes import Axes
+from matplotlib.figure import Figure
 from matplotlib.lines import Line2D
+import matplotlib.pyplot as plt
 
-__all__ = ("add_custom_legend",)
+from ..model import lfric, um
+from ..model.base import Model
+from ..coord import get_cube_rel_days
+
+__all__ = (
+    "add_custom_legend",
+    "capitalise",
+    "hcross",
+    "make_list_2d",
+    "map_scatter",
+    "timeseries_1d",
+    "timeseries_2d",
+)
 
 
-def add_custom_legend(ax_or_fig, styles_and_labels, **leg_kw):
+def add_custom_legend(ax_or_fig: Union[Axes, Figure], styles_and_labels: dict, **leg_kw) -> None:
     """
     Add a custom legend to a matplotlib axis or figure.
 
@@ -35,3 +53,87 @@ def add_custom_legend(ax_or_fig, styles_and_labels, **leg_kw):
             ax_or_fig.add_artist(leg)
     except AttributeError:
         pass
+
+
+def capitalise(s: str, sep_old: Optional[str] = "_", sep_new: Optional[str] = " ") -> str:
+    """Split the string and capitalise each word."""
+    return sep_new.join([i.capitalize() for i in s.split(sep_old)])
+
+
+def hcross(
+    cube: Cube, ax: Optional[Axes] = None, model: Optional[Model] = um, **kw_plt
+) -> Optional[Axes]:
+    """Plot a horizontal cross-section aka lat-lon map of a 2D cube."""
+    newax = False
+    if ax is None:
+        ax = plt.axes()
+        newax = True
+    fig = ax.figure
+    lons = cube.coord(um.x).points
+    lats = cube.coord(um.y).points
+    mappable = ax.pcolormesh(lons, lats, cube.data, **kw_plt)
+    fig.colorbar(mappable, ax=ax)
+    if newax:
+        return ax
+
+
+def make_list_2d(
+    list_x: Sequence[str], list_y: Sequence[str], transpose: Optional[bool] = False
+) -> list:
+    """Create a nested list out of 2 given lists."""
+    if transpose:
+        return [[f"{key_y}-{key_x}" for key_x in list_x] for key_y in list_y]
+    else:
+        return [[f"{key_x}-{key_y}" for key_x in list_x] for key_y in list_y]
+
+
+def map_scatter(
+    cube: Cube, ax: Optional[Axes] = None, model: Optional[Model] = lfric, **kw_plt
+) -> Optional[Axes]:
+    """Plot a lat-lon scatter plot of a 2D cube."""
+    newax = False
+    if ax is None:
+        ax = plt.axes()
+        newax = True
+    fig = ax.figure
+    # This doesn't work because lons and lats in the mesh are of size N+2
+    # while the data array is of size N
+    # lons, lats = cube.mesh.node_coords
+    # lons, lats = lons.points, lats.points
+    lons = cube.coord(model.x).points
+    lats = cube.coord(model.y).points
+    mappable = ax.scatter(lons, lats, c=cube.data, **kw_plt)
+    fig.colorbar(mappable, ax=ax)
+    if newax:
+        return ax
+
+
+def timeseries_1d(
+    cube: Cube, ax: Optional[Axes] = None, model: Optional[Model] = um, **kw_plt
+) -> Optional[Axes]:
+    """Plot time series of a 1D cube."""
+    newax = False
+    if ax is None:
+        ax = plt.axes()
+        newax = True
+    days = get_cube_rel_days(cube, model=um)
+    ax.plot(days, cube.data, **kw_plt)
+    if newax:
+        return ax
+
+
+def timeseries_2d(
+    cube: Cube, ax: Optional[Axes] = None, model: Optional[Model] = um, **kw_plt
+) -> Optional[Axes]:
+    """Plot time series of a 2D cube."""
+    newax = False
+    if ax is None:
+        ax = plt.axes()
+        newax = True
+    fig = ax.figure
+    days = get_cube_rel_days(cube, model=um)
+    z = cube.coord(um.z).points
+    mappable = ax.pcolormesh(days, z, cube.data.T, **kw_plt)
+    fig.colorbar(mappable, ax=ax)
+    if newax:
+        return ax