+++ src/aeolus/runtime.py
@@ -7,9 +7,10 @@ __all__ = ("RUNTIME",)
 class RuntimeOpts(threading.local):
     """Run-time configuration controller."""
 
-    def __init__(self, figsave_stamp=False):
+    def __init__(self, figsave_stamp=False, figsave_reduce_size=True):
         """Initialise run-time configuration."""
         self.__dict__["figsave_stamp"] = figsave_stamp
+        self.__dict__["figsave_reduce_size"] = figsave_reduce_size
 
     def __repr__(self):
         """Repr of run-time configuration."""