+++ docs/source/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 
 * Add `lfric` functions in a submodule with the same name.
 * Add `esmf_regrid` as a dependency.
+* Add various plotting utilities.
 * Update dependencies.